% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{Query Data Commons Using SPARQL}
\usage{
query(query_string)
}
\arguments{
\item{query_string}{required, SPARQL query string.}
}
\value{
A populated data frame with columns specified by the SPARQL query.
}
\description{
This function allows you to build R dataframes using data from
the Data Commons Open Knowledge Graph via SPARQL queries to
the Data Commons Query API.
}
\examples{
query_string = "SELECT ?pop ?Unemployment
    WHERE {
      ?pop typeOf StatisticalPopulation .
      ?o typeOf Observation .
      ?pop dcid (\\"dc/p/qep2q2lcc3rcc\\" \\"dc/p/gmw3cn8tmsnth\\" \\"dc/p/92cxc027krdcd\\") .
      ?o observedNode ?pop .
      ?o measuredValue ?Unemployment
    }"
df = query(query_string)
}
