% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{set_api_key}
\alias{set_api_key}
\title{Set your API Key}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{(string) your API Key.}
}
\description{
Using the Data Commons API requires you to provision an API key on
GCP and enabling th Data Commons API. See
\url{http://docs.datacommons.org/api/setup.html}
}
\examples{
# Say you got an API Key: 12345678, in the R console:
set_api_key("12345678")
}
